--- a/archiver_options.go
+++ b/archiver_options.go
@@ -12,6 +12,7 @@ type ArchiverOption func(*archiverOptions) error
 type archiverOptions struct {
 	method      uint16
 	concurrency int
+	bufferSize  int
 	stageDir    string
 	offset      int64
 }
@@ -25,13 +26,28 @@ func WithArchiverMethod(method uint16) ArchiverOption {
 }
 
 // WithArchiverConcurrency will set the maximum number of files to be
-// compressed concurrently. The default is GOMAXPROCS.
+// compressed concurrently. The default is set to runtime.NumCPU().
 func WithArchiverConcurrency(n int) ArchiverOption {
 	return func(o *archiverOptions) error {
 		if n <= 0 {
 			return ErrMinConcurrency
 		}
 		o.concurrency = n
+		return nil
+	}
+}
+
+// WithArchiverBufferSize sets the buffer size for each file to be compressed
+// concurrently. If a compressed file's data exceeds the buffer size, a
+// temporary file is written (to the stage directory) to hold the additional
+// data. The default is 2 mebibytes, so if concurrency is 16, 32 mebibytes of
+// memory will be allocated.
+//
+// If set to -1, no buffer will be used and all compressed file content will be
+// written to temporary files before being written back to the zip file.
+func WithArchiverBufferSize(n int) ArchiverOption {
+	return func(o *archiverOptions) error {
+		o.bufferSize = n
 		return nil
 	}
 }