--- a/internal/filepool/filepool_test.go
+++ b/internal/filepool/filepool_test.go
@@ -2,6 +2,7 @@ package filepool
 
 import (
 	"bytes"
+	"errors"
 	"fmt"
 	"io/ioutil"
 	"os"
@@ -14,40 +15,60 @@ import (
 )
 
 func TestFilePoolSizes(t *testing.T) {
-	dir, err := ioutil.TempDir("", "fastzip-filepool")
-	require.NoError(t, err)
+	tests := []struct {
+		size int
+		err  error
+	}{
+		{-1, ErrPoolSizeLessThanZero},
+		{0, ErrPoolSizeLessThanZero},
+		{4, nil},
+		{8, nil},
+	}
 
-	for i := 0; i < 16; i++ {
-		fp, err := New(dir, i)
-		if i == 0 {
-			require.Error(t, err, "size of zero should return error")
-			continue
-		}
-		require.NoError(t, err)
+	for _, tc := range tests {
+		t.Run(fmt.Sprintf("size %d", tc.size), func(t *testing.T) {
+			dir, err := ioutil.TempDir("", "fastzip-filepool")
+			require.NoError(t, err)
+			defer os.RemoveAll(dir)
 
-		for n := 0; n < i; n++ {
-			_, err = os.Lstat(filepath.Join(dir, fmt.Sprintf("fastzip_%02d", n)))
-			assert.NoError(t, err, fmt.Sprintf("fastzip_%02d should exist", n))
-		}
+			fp, err := New(dir, tc.size, -1)
+			require.Equal(t, tc.err, err)
+			if tc.err != nil {
+				return
+			}
 
-		assert.NoError(t, fp.Close())
+			// writing should produce the temporary file
+			for i := 0; i < tc.size; i++ {
+				f := fp.Get()
+				_, err = f.Write([]byte("foobar"))
+				assert.NoError(t, err)
+				fp.Put(f)
 
-		for n := 0; n < i; n++ {
-			_, err = os.Lstat(filepath.Join(dir, fmt.Sprintf("fastzip_%02d", n)))
-			assert.Error(t, err, fmt.Sprintf("fastzip_%02d shouldn't exist", n))
-		}
+				_, err = os.Lstat(filepath.Join(dir, fmt.Sprintf("fastzip_%02d", i)))
+				assert.NoError(t, err, fmt.Sprintf("fastzip_%02d should exist", i))
+			}
+
+			// closing should cleanup temporary files
+			assert.NoError(t, fp.Close())
+			for i := 0; i < tc.size; i++ {
+				_, err = os.Lstat(filepath.Join(dir, fmt.Sprintf("fastzip_%02d", i)))
+				assert.Error(t, err, fmt.Sprintf("fastzip_%02d shouldn't exist", i))
+			}
+		})
 	}
 }
 
 func TestFilePoolReset(t *testing.T) {
 	dir, err := ioutil.TempDir("", "fastzip-filepool")
 	require.NoError(t, err)
+	defer os.RemoveAll(dir)
 
-	fp, err := New(dir, 16)
+	fp, err := New(dir, 16, -1)
 	require.NoError(t, err)
-	for i := 0; i < 16; i++ {
+	for i := range fp.files {
 		file := fp.Get()
-		file.Write(bytes.Repeat([]byte("0"), i))
+		_, err = file.Write(bytes.Repeat([]byte("0"), i))
+		assert.NoError(t, err)
 
 		b, err := ioutil.ReadAll(file)
 		assert.NoError(t, err)
@@ -61,7 +82,7 @@ func TestFilePoolReset(t *testing.T) {
 		fp.Put(file)
 	}
 
-	for i := 0; i < 16; i++ {
+	for range fp.files {
 		file := fp.Get()
 
 		b, err := ioutil.ReadAll(file)
@@ -79,26 +100,29 @@ func TestFilePoolReset(t *testing.T) {
 func TestFilePoolCloseError(t *testing.T) {
 	dir, err := ioutil.TempDir("", "fastzip-filepool")
 	require.NoError(t, err)
+	defer os.RemoveAll(dir)
 
-	fp, err := New(dir, 16)
+	fp, err := New(dir, 16, -1)
 	require.NoError(t, err)
 
 	for _, file := range fp.files {
+		f := fp.Get()
+		_, err := f.Write([]byte("foobar"))
+		assert.NoError(t, err)
+		fp.Put(f)
+
 		require.NoError(t, file.f.Close())
 	}
 
 	err = fp.Close()
 	require.Error(t, err, "expected already closed error")
+	assert.Contains(t, err.Error(), "file already closed\n")
 	count := 0
 	for {
-		u, ok := err.(interface {
-			Unwrap() error
-		})
-		if !ok {
+		count++
+		if err = errors.Unwrap(err); err == nil {
 			break
 		}
-		err = u.Unwrap()
-		count++
 	}
 	assert.Equal(t, 16, count)
 }
@@ -111,11 +135,79 @@ func TestFilePoolNoErrorOnAlreadyDeleted(t *testing.T) {
 	dir, err := ioutil.TempDir("", "fastzip-filepool")
 	require.NoError(t, err)
 
-	fp, err := New(dir, 16)
+	fp, err := New(dir, 16, -1)
 	require.NoError(t, err)
+
+	for range fp.files {
+		f := fp.Get()
+		_, err := f.Write([]byte("foobar"))
+		assert.NoError(t, err)
+		fp.Put(f)
+	}
 
 	err = os.RemoveAll(dir)
 	require.NoError(t, err)
 
 	assert.NoError(t, fp.Close())
+}
+
+func TestFilePoolFileBuffer(t *testing.T) {
+	dir, err := ioutil.TempDir("", "fastzip-filepool")
+	require.NoError(t, err)
+	defer os.RemoveAll(dir)
+
+	tests := map[string]struct {
+		data       []byte
+		fileExists bool
+	}{
+		"below buffer length": {
+			data:       []byte("123456789"),
+			fileExists: false,
+		},
+		"equal to buffer length": {
+			data:       []byte("1234567890"),
+			fileExists: false,
+		},
+		"above buffer length": {
+			data:       []byte("1234567890x"),
+			fileExists: true,
+		},
+	}
+
+	for tn, tc := range tests {
+		t.Run(tn, func(t *testing.T) {
+			fp, err := New(dir, 1, 10)
+			require.NoError(t, err)
+			defer fp.Close()
+			require.Len(t, fp.files, 1)
+
+			f := fp.files[0]
+			n, err := f.Write(tc.data)
+			assert.NoError(t, err)
+			assert.Equal(t, len(tc.data), n)
+
+			_, err = os.Lstat(filepath.Join(dir, "fastzip_00"))
+			if tc.fileExists {
+				assert.NoError(t, err, "fastzip_00 should exist")
+			} else {
+				assert.Error(t, err, "fastzip_00 should not exist")
+			}
+
+			// split reads to ensure read/write indexes track correctly
+			buf := make([]byte, 20)
+			size := 0
+			{
+				n, err := f.Read(buf[:5])
+				assert.NoError(t, err)
+				size += n
+			}
+			{
+				n, err := f.Read(buf[5:])
+				assert.NoError(t, err)
+				size += n
+			}
+
+			assert.Equal(t, tc.data, buf[:size])
+		})
+	}
 }